+++ test/wal-windows-test.el
@@ -28,12 +28,6 @@
 
     (was-called-with delete-other-windows (list 'window))))
 
-(ert-deftest test-wal/popper-echo-transform ()
-  (should (string-match "variable: testing" (wal/popper-echo-transform "*helpful variable: testing*")))
-  (should (string-match "up: testing" (wal/popper-echo-transform "* docker-compose up --rm testing*")))
-  (should (string-match "logs: is_testing_2" (wal/popper-echo-transform "* docker container logs --tail 10 is_testing_2 *")))
-  (should (string-match "*unchanged*" (wal/popper-echo-transform "*unchanged*"))))
-
 (ert-deftest test-wal/popper--spared-p ()
   (let ((wal/spared-popups '("testing")))
 
@@ -50,10 +44,13 @@
 
     (with-mock ((buffer-live-p . #'always)
                 kill-some-buffers
+                kill-buffer
                 (buffer-name . #'wal/rf))
 
       (wal/kill-some-popups)
+      (wal/kill-some-popups t)
 
-      (was-called-with kill-some-buffers (list (list "dying"))))))
+      (was-called-with kill-some-buffers (list (list "dying")))
+      (was-called-with kill-buffer (list "dying")))))
 
 ;;; wal-windows-test.el ends here